requirements.txt
imagehash
pillow
flask

